+++ OpenKh.Patcher/BarEntryExtractor/IfApplyToBarEntryDelegate.cs
@@ -1,0 +1,11 @@
+using OpenKh.Kh2;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace OpenKh.Patcher.BarEntryExtractor
+{
+    public delegate bool IfApplyToBarEntryDelegate(string EntryName, Bar.EntryType EntryType, int EntryIndex);
+}
