+++ OpenKh.Tools.ModsManager/Views/YamlGeneratorVM.cs
@@ -3,11 +3,13 @@ using OpenKh.Common;
 using OpenKh.Kh2;
 using OpenKh.Kh2.Messages;
 using OpenKh.Patcher;
+using OpenKh.Patcher.BarEntryExtractor;
 using OpenKh.Tools.Common.Wpf;
 using OpenKh.Tools.ModsManager.Models.ViewHelper;
 using OpenKh.Tools.ModsManager.Services;
 using System;
 using System.Collections.Generic;
+using System.Collections.Immutable;
 using System.ComponentModel;
 using System.IO;
 using System.Linq;
@@ -169,576 +171,15 @@ namespace OpenKh.Tools.ModsManager.Views
         private readonly KeywordsMatcherService _keywordsMatcherService = new KeywordsMatcherService();
         private readonly ISerializer _listSer = new SerializerBuilder()
             .Build();
+        private readonly ProvideExtractorsService _provideExtractorsService = new ProvideExtractorsService();
 
         private record MixedSource(CopySourceFile CopySourceFile, AssetFile AssetFile);
 
-        private delegate bool IfApplyToBarEntryDelegate(string EntryName, Bar.EntryType EntryType, int EntryIndex);
-
-        private record SourceBuilderArg(string DestName, string DestType, string SourceName, string OriginalRelativePath);
-
-        private record Extractor(
-            IfApplyToBarEntryDelegate IfApply,
-            Func<string, bool> SourceFileTest,
-            string FileExtension,
-            Func<Bar.Entry, Task<byte[]>> ExtractAsync,
-            Func<SourceBuilderArg, IEnumerable<AssetFile>> SourceBuilder = null
-        );
-
         public YamlGeneratorVM()
         {
-            var extractors = new List<Extractor>();
-
-
-            extractors.Add(new Extractor(
-                IfApply: (name, type, index) => name == "trsr" && type == Bar.EntryType.List,
-                SourceFileTest: (relativePath) => relativePath == "03system.bin",
-                FileExtension: ".yml",
-                ExtractAsync: async (Bar.Entry barEntry) =>
-                {
-                    await Task.Yield();
-                    return Encoding.UTF8.GetBytes(
-                        _listSer.Serialize(
-                            OpenKh.Kh2.SystemData.Trsr.Read(barEntry.Stream.FromBegin())
-                        )
-                    );
-                },
-                SourceBuilder: arg => CreateSourceFromArgs(
-                    new AssetFile
-                    {
-                        Name = "trsr",
-                        Type = "list",
-                        Method = "listpatch",
-                        Source = CreateSourceFromArgs(
-                            new AssetFile
-                            {
-                                Name = arg.SourceName,
-                                Type = "trsr",
-                            }
-                        ),
-                    }
-                )
-            ));
-
-            extractors.Add(new Extractor(
-                IfApply: (name, type, index) => name == "item" && type == Bar.EntryType.List,
-                SourceFileTest: (relativePath) => relativePath == "03system.bin",
-                FileExtension: ".yml",
-                ExtractAsync: async (Bar.Entry barEntry) =>
-                {
-                    await Task.Yield();
-                    return Encoding.UTF8.GetBytes(
-                        _listSer.Serialize(
-                            OpenKh.Kh2.SystemData.Item.Read(barEntry.Stream.FromBegin())
-                        )
-                    );
-                },
-                SourceBuilder: arg => CreateSourceFromArgs(
-                    new AssetFile
-                    {
-                        Name = "item",
-                        Type = "list",
-                        Method = "listpatch",
-                        Source = CreateSourceFromArgs(
-                            new AssetFile
-                            {
-                                Name = arg.SourceName,
-                                Type = "item",
-                            }
-                        ),
-                    }
-                )
-            ));
-
-            extractors.Add(new Extractor(
-                IfApply: (name, type, index) => name == "fmlv" && type == Bar.EntryType.List,
-                SourceFileTest: (relativePath) => relativePath == "00battle.bin",
-                FileExtension: ".yml",
-                ExtractAsync: async (Bar.Entry barEntry) =>
-                {
-                    await Task.Yield();
-                    return Encoding.UTF8.GetBytes(
-                        _listSer.Serialize(
-                            OpenKh.Kh2.Battle.Fmlv.Read(barEntry.Stream.FromBegin())
-                        )
-                    );
-                },
-                SourceBuilder: arg => CreateSourceFromArgs(
-                    new AssetFile
-                    {
-                        Name = "fmlv",
-                        Type = "list",
-                        Method = "listpatch",
-                        Source = CreateSourceFromArgs(
-                            new AssetFile
-                            {
-                                Name = arg.SourceName,
-                                Type = "fmlv",
-                            }
-                        ),
-                    }
-                )
-            ));
-
-            extractors.Add(new Extractor(
-                IfApply: (name, type, index) => name == "lvup" && type == Bar.EntryType.List,
-                SourceFileTest: (relativePath) => relativePath == "00battle.bin",
-                FileExtension: ".yml",
-                ExtractAsync: async (Bar.Entry barEntry) =>
-                {
-                    await Task.Yield();
-                    return Encoding.UTF8.GetBytes(
-                        _listSer.Serialize(
-                            OpenKh.Kh2.Battle.Lvup.Read(barEntry.Stream.FromBegin())
-                        )
-                    );
-                },
-                SourceBuilder: arg => CreateSourceFromArgs(
-                    new AssetFile
-                    {
-                        Name = "lvup",
-                        Type = "list",
-                        Method = "listpatch",
-                        Source = CreateSourceFromArgs(
-                            new AssetFile
-                            {
-                                Name = arg.SourceName,
-                                Type = "lvup",
-                            }
-                        ),
-                    }
-                )
-            ));
-
-            extractors.Add(new Extractor(
-                IfApply: (name, type, index) => name == "bons" && type == Bar.EntryType.List,
-                SourceFileTest: (relativePath) => relativePath == "00battle.bin",
-                FileExtension: ".yml",
-                ExtractAsync: async (Bar.Entry barEntry) =>
-                {
-                    await Task.Yield();
-                    return Encoding.UTF8.GetBytes(
-                        _listSer.Serialize(
-                            OpenKh.Kh2.Battle.Bons.Read(barEntry.Stream.FromBegin())
-                        )
-                    );
-                },
-                SourceBuilder: arg => CreateSourceFromArgs(
-                    new AssetFile
-                    {
-                        Name = "bons",
-                        Type = "list",
-                        Method = "listpatch",
-                        Source = CreateSourceFromArgs(
-                            new AssetFile
-                            {
-                                Name = arg.SourceName,
-                                Type = "bons",
-                            }
-                        ),
-                    }
-                )
-            ));
-
-            extractors.Add(new Extractor(
-                IfApply: (name, type, index) => name == "atkp" && type == Bar.EntryType.List,
-                SourceFileTest: (relativePath) => relativePath == "00battle.bin",
-                FileExtension: ".yml",
-                ExtractAsync: async (Bar.Entry barEntry) =>
-                {
-                    await Task.Yield();
-                    return Encoding.UTF8.GetBytes(
-                        _listSer.Serialize(
-                            OpenKh.Kh2.Battle.Atkp.Read(barEntry.Stream.FromBegin())
-                        )
-                    );
-                },
-                SourceBuilder: arg => CreateSourceFromArgs(
-                    new AssetFile
-                    {
-                        Name = "atkp",
-                        Type = "list",
-                        Method = "listpatch",
-                        Source = CreateSourceFromArgs(
-                            new AssetFile
-                            {
-                                Name = arg.SourceName,
-                                Type = "atkp",
-                            }
-                        ),
-                    }
-                )
-            ));
-
-            extractors.Add(new Extractor(
-                IfApply: (name, type, index) => name == "plrp" && type == Bar.EntryType.List,
-                SourceFileTest: (relativePath) => relativePath == "00battle.bin",
-                FileExtension: ".yml",
-                ExtractAsync: async (Bar.Entry barEntry) =>
-                {
-                    await Task.Yield();
-                    return Encoding.UTF8.GetBytes(
-                        _listSer.Serialize(
-                            OpenKh.Kh2.Battle.Plrp.Read(barEntry.Stream.FromBegin())
-                        )
-                    );
-                },
-                SourceBuilder: arg => CreateSourceFromArgs(
-                    new AssetFile
-                    {
-                        Name = "plrp",
-                        Type = "list",
-                        Method = "listpatch",
-                        Source = CreateSourceFromArgs(
-                            new AssetFile
-                            {
-                                Name = arg.SourceName,
-                                Type = "plrp",
-                            }
-                        ),
-                    }
-                )
-            ));
-
-            extractors.Add(new Extractor(
-                IfApply: (name, type, index) => name == "cmd" && type == Bar.EntryType.List,
-                SourceFileTest: (relativePath) => relativePath == "03system.bin",
-                FileExtension: ".yml",
-                ExtractAsync: async (Bar.Entry barEntry) =>
-                {
-                    await Task.Yield();
-                    return Encoding.UTF8.GetBytes(
-                        _listSer.Serialize(
-                            OpenKh.Kh2.SystemData.Cmd.Read(barEntry.Stream.FromBegin())
-                        )
-                    );
-                },
-                SourceBuilder: arg => CreateSourceFromArgs(
-                    new AssetFile
-                    {
-                        Name = "cmd",
-                        Type = "list",
-                        Method = "listpatch",
-                        Source = CreateSourceFromArgs(
-                            new AssetFile
-                            {
-                                Name = arg.SourceName,
-                                Type = "cmd",
-                            }
-                        ),
-                    }
-                )
-            ));
-
-            extractors.Add(new Extractor(
-                IfApply: (name, type, index) => name == "enmp" && type == Bar.EntryType.List,
-                SourceFileTest: (relativePath) => relativePath == "00battle.bin",
-                FileExtension: ".yml",
-                ExtractAsync: async (Bar.Entry barEntry) =>
-                {
-                    await Task.Yield();
-                    return Encoding.UTF8.GetBytes(
-                        _listSer.Serialize(
-                            OpenKh.Kh2.Battle.Enmp.Read(barEntry.Stream.FromBegin())
-                        )
-                    );
-                },
-                SourceBuilder: arg => CreateSourceFromArgs(
-                    new AssetFile
-                    {
-                        Name = "enmp",
-                        Type = "list",
-                        Method = "listpatch",
-                        Source = CreateSourceFromArgs(
-                            new AssetFile
-                            {
-                                Name = arg.SourceName,
-                                Type = "enmp",
-                            }
-                        ),
-                    }
-                )
-            ));
-
-            extractors.Add(new Extractor(
-                IfApply: (name, type, index) => name == "sklt" && type == Bar.EntryType.List,
-                SourceFileTest: (relativePath) => relativePath == "03system.bin",
-                FileExtension: ".yml",
-                ExtractAsync: async (Bar.Entry barEntry) =>
-                {
-                    await Task.Yield();
-                    return Encoding.UTF8.GetBytes(
-                        _listSer.Serialize(
-                            OpenKh.Kh2.SystemData.Sklt.Read(barEntry.Stream.FromBegin())
-                        )
-                    );
-                },
-                SourceBuilder: arg => CreateSourceFromArgs(
-                    new AssetFile
-                    {
-                        Name = "sklt",
-                        Type = "list",
-                        Method = "listpatch",
-                        Source = CreateSourceFromArgs(
-                            new AssetFile
-                            {
-                                Name = arg.SourceName,
-                                Type = "sklt",
-                            }
-                        ),
-                    }
-                )
-            ));
-
-            extractors.Add(new Extractor(
-                IfApply: (name, type, index) => name == "przt" && type == Bar.EntryType.List,
-                SourceFileTest: (relativePath) => relativePath == "00battle.bin",
-                FileExtension: ".yml",
-                ExtractAsync: async (Bar.Entry barEntry) =>
-                {
-                    await Task.Yield();
-                    return Encoding.UTF8.GetBytes(
-                        _listSer.Serialize(
-                            OpenKh.Kh2.Battle.Przt.Read(barEntry.Stream.FromBegin())
-                        )
-                    );
-                },
-                SourceBuilder: arg => CreateSourceFromArgs(
-                    new AssetFile
-                    {
-                        Name = "przt",
-                        Type = "list",
-                        Method = "listpatch",
-                        Source = CreateSourceFromArgs(
-                            new AssetFile
-                            {
-                                Name = arg.SourceName,
-                                Type = "przt",
-                            }
-                        ),
-                    }
-                )
-            ));
-
-            extractors.Add(new Extractor(
-                IfApply: (name, type, index) => name == "magc" && type == Bar.EntryType.List,
-                SourceFileTest: (relativePath) => relativePath == "00battle.bin",
-                FileExtension: ".yml",
-                ExtractAsync: async (Bar.Entry barEntry) =>
-                {
-                    await Task.Yield();
-                    return Encoding.UTF8.GetBytes(
-                        _listSer.Serialize(
-                            OpenKh.Kh2.Battle.Magc.Read(barEntry.Stream.FromBegin())
-                        )
-                    );
-                },
-                SourceBuilder: arg => CreateSourceFromArgs(
-                    new AssetFile
-                    {
-                        Name = "magc",
-                        Type = "list",
-                        Method = "listpatch",
-                        Source = CreateSourceFromArgs(
-                            new AssetFile
-                            {
-                                Name = arg.SourceName,
-                                Type = "magc",
-                            }
-                        ),
-                    }
-                )
-            ));
-
-            extractors.Add(new Extractor(
-                IfApply: (name, type, index) => type == Bar.EntryType.Model,
-                SourceFileTest: (relativePath) => true,
-                FileExtension: ".model",
-                ExtractAsync: async (Bar.Entry barEntry) =>
-                {
-                    await Task.Yield();
-                    return barEntry.Stream.ReadAllBytes();
-                }
-            ));
-
-            extractors.Add(new Extractor(
-                IfApply: (name, type, index) => type == Bar.EntryType.ModelTexture,
-                SourceFileTest: (relativePath) => true,
-                FileExtension: ".tim",
-                ExtractAsync: async (Bar.Entry barEntry) =>
-                {
-                    await Task.Yield();
-                    return barEntry.Stream.ReadAllBytes();
-                }
-            ));
-
-
-            Extractor CreateMessageExtractor(
-                Regex sourceFilePattern,
-                Func<Match, string> languageSelector,
-                IMessageDecode decoder
-            )
-            {
-                return new Extractor(
-                    IfApply: (name, type, index) => type == Bar.EntryType.List,
-                    SourceFileTest: relativePath => sourceFilePattern.IsMatch(relativePath),
-                    FileExtension: ".yml",
-                    ExtractAsync: async (Bar.Entry barEntry) =>
-                    {
-                        await Task.Yield();
-                        var msgEntries = Msg.Read(barEntry.Stream.FromBegin());
-                        return Encoding.UTF8.GetBytes(
-                            _listSer.Serialize(
-                                msgEntries
-                                    .Select(
-                                        msgEntry => new IdText(
-                                            msgEntry.Id,
-                                            MsgSerializer.SerializeText(decoder.Decode(msgEntry.Data))
-                                        )
-                                    )
-                            )
-                        );
-                    },
-                    SourceBuilder: arg => CreateSourceFromArgs(
-                        new AssetFile
-                        {
-                            Name = "sys",
-                            Type = "list",
-                            Method = "kh2msg",
-                            Source = CreateSourceFromArgs(
-                                new AssetFile
-                                {
-                                    Name = "sys",
-                                    Language = languageSelector(sourceFilePattern.Match(arg.OriginalRelativePath)),
-                                }
-                            ),
-                        }
-                    )
-                );
-            }
-
-            extractors.Add(CreateMessageExtractor(
-                sourceFilePattern: new Regex(@"msg/(?<language>us|fr|gr|it|sp)/(sys|[a-z]{2})\.bar"),
-                languageSelector: match => match.Groups["language"].Value,
-                decoder: Encoders.InternationalSystem
-            ));
-
-            extractors.Add(CreateMessageExtractor(
-                sourceFilePattern: new Regex(@"msg/tr/sys\.bar"),
-                languageSelector: match => "tr",
-                decoder: Encoders.TurkishSystem
-            ));
-            extractors.Add(CreateMessageExtractor(
-                sourceFilePattern: new Regex(@"msg/tr/[a-z]{2}\.bar"),
-                languageSelector: match => "tr",
-                decoder: Encoders.TurkishSystem
-            ));
-
-            extractors.Add(CreateMessageExtractor(
-                sourceFilePattern: new Regex(@"msg/jp/sys\.bar"),
-                languageSelector: match => "jp",
-                decoder: Encoders.JapaneseSystem
-            ));
-            extractors.Add(CreateMessageExtractor(
-                sourceFilePattern: new Regex(@"msg/jp/[a-z]{2}.bar"),
-                languageSelector: match => "je",
-                decoder: Encoders.JapaneseEvent
-            ));
-
-
-            extractors.Add(new Extractor(
-                IfApply: (name, type, index) => type == Bar.EntryType.Bdx,
-                SourceFileTest: (relativePath) => true,
-                FileExtension: ".bdscript",
-                ExtractAsync: async (Bar.Entry barEntry) =>
-                {
-                    await Task.Yield();
-                    var decoder = new BdxDecoder(
-                        read: new MemoryStream(barEntry.Stream.ReadAllBytes(), false),
-                        codeRevealer: true,
-                        codeRevealerLabeling: true
-                    );
-                    return Encoding.UTF8.GetBytes(
-                        BdxDecoder.TextFormatter.Format(decoder)
-                    );
-                },
-                SourceBuilder: arg => CreateSourceFromArgs(
-                    new AssetFile
-                    {
-                        Name = arg.DestName,
-                        Type = "bdx",
-                        Method = "bdscript",
-                        Source = CreateSourceFromArgs(
-                            new AssetFile
-                            {
-                                Name = arg.SourceName,
-                            }
-                        ),
-                    }
-                )
-            ));
-
-
-            extractors.Add(new Extractor(
-                IfApply: (name, type, index) => type == Bar.EntryType.AreaDataScript,
-                SourceFileTest: (relativePath) => true,
-                FileExtension: ".script",
-                ExtractAsync: async (Bar.Entry barEntry) =>
-                {
-                    await Task.Yield();
-                    return Encoding.UTF8.GetBytes(
-                        Kh2.Ard.AreaDataScript.Decompile(
-                            Kh2.Ard.AreaDataScript.Read(barEntry.Stream.FromBegin())
-                        )
-                    );
-                },
-                SourceBuilder: arg => CreateSourceFromArgs(
-                    new AssetFile
-                    {
-                        Name = arg.DestName,
-                        Type = "areadatascript",
-                        Method = "areadatascript",
-                        Source = CreateSourceFromArgs(
-                            new AssetFile
-                            {
-                                Name = arg.SourceName,
-                            }
-                        ),
-                    }
-                )
-            ));
-
-
-            extractors.Add(new Extractor(
-                IfApply: (name, type, index) => type == Bar.EntryType.AreaDataSpawn,
-                SourceFileTest: (relativePath) => true,
-                FileExtension: ".yml",
-                ExtractAsync: async (Bar.Entry barEntry) =>
-                {
-                    await Task.Yield();
-                    return Encoding.UTF8.GetBytes(
-                        _listSer.Serialize(
-                            Kh2.Ard.SpawnPoint.Read(
-                                barEntry.Stream.FromBegin()
-                            )
-                        )
-                    );
-                },
-                SourceBuilder: arg => CreateSourceFromArgs(
-                    new AssetFile
-                    {
-                        Name = arg.DestName,
-                        Type = "areadataspawn",
-                        Method = "spawnpoint",
-                        Source = CreateSourceFromArgs(
-                            new AssetFile
-                            {
-                                Name = arg.SourceName,
-                            }
-                        ),
-                    }
-                )
-            ));
-
+            var extractors = _provideExtractorsService
+                .GetExtractors()
+                .ToImmutableArray();
 
             GetDiffService diffTool = null;
 
@@ -1412,7 +853,5 @@ namespace OpenKh.Tools.ModsManager.Views
                 LoadPref(pref);
             }
         }
-
-        private record IdText(int Id, string Text);
     }
 }
